+++ kaldb/src/main/java/com/slack/kaldb/histogram/Histogram.java
@@ -15,5 +15,5 @@ public interface Histogram {
   List<HistogramBucket> getBuckets();
 
   // Return count of number of values in histogram.
-  long count();
+  double count();
 }
